organization of shell function and their helpers in two different files + implementation of exit + fix ctrl D
